Restructuring and fixing errors
* Activate medication works but contains bugs
* Deactivate medication works
* Delete medication works
