+++ src/models/Metadata.js
@@ -2,9 +2,11 @@ import mongoose from "mongoose";
 
 const Schema = mongoose.Schema;
 
-const MetadataSchema = new Schema({
+export const MetadataSchema = new Schema({
     name: String,
     lastScheduledAt: Date,
 });
 
-export const Metadata = mongoose.model("Metadata", MetadataSchema);
+const Metadata = mongoose.model("Metadata", MetadataSchema);
+
+export default Metadata;
