+++ src/models/Occurrence.js
@@ -23,23 +23,6 @@ export const OccurrenceSchema = new Schema({
     },
 });
 
-export const OccurrenceGroupSchema = new Schema({
-    occurrences: [
-        {
-            type: Schema.Types.ObjectId,
-            ref: "Occurrence",
-        },
-    ],
-    scheduledDate: {
-        type: Date,
-    },
-    user: {
-        type: Schema.Types.ObjectId,
-        ref: "User",
-    },
-    emergencyJobId: {
-        //should we be using Schema.Types.ObjectId???
-        type: Schema.Types.ObjectId,
-        default: null,
-    },
-});
+const Occurrence = mongoose.model("Occurrence", OccurrenceSchema);
+
+export default Occurrence;
