+++ src/models/User.js
@@ -1,34 +1,8 @@
 import bcrypt from "bcrypt";
 import mongoose from "mongoose";
+import { SettingsSchema } from "./Settings";
 
 const Schema = mongoose.Schema;
-
-export const SettingsSchema = new Schema({
-    notificationSettings: {
-        silenceAll: {
-            type: Boolean,
-            default: false,
-        },
-        hideMedName: {
-            type: Boolean,
-            default: true,
-        },
-    },
-    hasCaregiverContact: {
-        type: Boolean,
-        default: false,
-    },
-    caregiverContact: {
-        name: {
-            type: String,
-        },
-        // phone numbers must be formatted as "+1*insert 10-digit phone number here*" in order to work
-        // with twilio's messaging service
-        phoneNumber: {
-            type: String,
-        },
-    },
-});
 
 export const UserSchema = new Schema({
     email: {
@@ -74,3 +48,6 @@ export const UserSchema = new Schema({
 UserSchema.methods.comparePassword = (password, hashPassword) => {
     return bcrypt.compareSync(password, hashPassword);
 };
+
+const User = mongoose.model("User", UserSchema);
+export default User;
